requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
billiard==4.2.0
# celery==5.3.6
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.5.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
fastapi-events==0.9.1
fastapi-pagination==0.12.13
greenlet==3.0.3
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.0.0
kombu==5.3.5
Mako==1.3.2
MarkupSafe==2.1.5
passlib==1.7.4
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
psycopg2
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.7
pytz==2024.1
redbird==0.7.1
requests==2.31.0
sentry-sdk==1.38.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.27.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.0.7
uvicorn==0.24.0.post1
vine==5.1.0
wcwidth==0.2.13
filetype==1.2.0
PyPDF2==2.0.0
python-docx
openai
tldextract

